requirements.txt
gradio
openai
google-genai
python-dotenv
mem0ai